+++ CHANGELOG.md
@@ -91,6 +91,7 @@
 - Accounting tool: added source filename of transaction records to cover page of PDF report.
 - Conversion tool: added new parser for Easy Crypto NZ exchange.
 - KuCoin parser: added new export formats for deposits and withdrawals.
+- BlockFi parser: added "Cc Rewards Redemption" transaction type.
 ### Changed
 - Conversion tool: openpyxl use read-only mode. ([#337](https://github.com/BittyTax/BittyTax/issues/337))
 - Accounting tool: openpyxl use read-only mode. ([#337](https://github.com/BittyTax/BittyTax/issues/337))